side_project/restful_api/requirements.txt
Flask
flask-restful
httpie
Flask-SQLAlchemy
Flask-Migrate
marshmallow
marshmallow-sqlalchemy
flask-marshmallow
psycopg2
